+++ engine/api/golang/kurtosis_engine_version/kurtosis_engine_version.go
@@ -1,0 +1,9 @@
+package kurtosis_engine_version
+
+const (
+	// !!!!!!!!!!!!!!!!!! DO NOT MODIFY THIS! IT WILL BE UPDATED AUTOMATICALLY DURING THE RELEASE PROCESS !!!!!!!!!!!!!!!
+	//This is necessary because we are comparing the version that is declared in the library with the running engine version
+	//before creating a new Kurtosis Context
+	KurtosisEngineVersion = "1.8.0"
+	// !!!!!!!!!!!!!!!!!! DO NOT MODIFY THIS! IT WILL BE UPDATED AUTOMATICALLY DURING THE RELEASE PROCESS !!!!!!!!!!!!!!!
+)