+++ engine/api/golang/lib/kurtosis_context/kurtosis_context.go
@@ -3,12 +3,16 @@ package kurtosis_context
 import (
 	"context"
 	"fmt"
+	"github.com/Masterminds/semver/v3"
 	"github.com/kurtosis-tech/kurtosis-core-api-lib/api/golang/kurtosis_core_rpc_api_bindings"
 	"github.com/kurtosis-tech/kurtosis-core-api-lib/api/golang/lib/enclaves"
 	"github.com/kurtosis-tech/kurtosis-engine-server/api/golang/kurtosis_engine_rpc_api_bindings"
+	"github.com/kurtosis-tech/kurtosis-engine-server/api/golang/kurtosis_engine_version"
 	"github.com/kurtosis-tech/stacktrace"
 	"github.com/sirupsen/logrus"
 	"google.golang.org/grpc"
+	"google.golang.org/grpc/codes"
+	"google.golang.org/grpc/status"
 	"google.golang.org/protobuf/types/known/emptypb"
 )
 
@@ -34,6 +38,7 @@ type KurtosisContext struct {
 // NewKurtosisContextFromLocalEngine
 // Attempts to create a KurtosisContext connected to a Kurtosis engine running locally
 func NewKurtosisContextFromLocalEngine() (*KurtosisContext, error) {
+	ctx := context.Background()
 	kurtosisEngineSocketStr := fmt.Sprintf("%v:%v", localHostIPAddressStr, DefaultKurtosisEngineServerPortNum)
 
 	// TODO SECURITY: Use HTTPS to ensure we're connecting to the real Kurtosis API servers
@@ -47,6 +52,41 @@ func NewKurtosisContextFromLocalEngine() (*KurtosisContext, error) {
 	}
 
 	engineServiceClient := kurtosis_engine_rpc_api_bindings.NewEngineServiceClient(conn)
+
+	getEngineInfoResponse, err := engineServiceClient.GetEngineInfo(ctx, &emptypb.Empty{})
+	if err != nil {
+		errorStr := "An error occurred getting engine info"
+		grpcErrorCode := status.Code(err)
+		if grpcErrorCode == codes.Unavailable {
+			errorStr = "The Kurtosis Engine Server is unavailable and is probably not running; you will need to start it using the Kurtosis CLI before you can create a connection to it"
+		}
+		return nil, stacktrace.Propagate(err, errorStr)
+	}
+	runningEngineVersionStr := getEngineInfoResponse.GetEngineVersion()
+
+	runningEngineSemver, err := semver.StrictNewVersion(runningEngineVersionStr)
+	if err != nil {
+		return nil, stacktrace.Propagate(err, "An error occurred parsing running engine version string '%v' to semantic version", runningEngineVersionStr)
+	}
+
+	runningEngineMajorVersion := runningEngineSemver.Major()
+	runningEngineMinorVersion := runningEngineSemver.Minor()
+
+	libraryEngineSemver, err := semver.StrictNewVersion(kurtosis_engine_version.KurtosisEngineVersion)
+	if err != nil {
+		return nil, stacktrace.Propagate(err, "An error occurred parsing library engine version string '%v' to semantic version", kurtosis_engine_version.KurtosisEngineVersion)
+	}
+	libraryEngineMajorVersion := libraryEngineSemver.Major()
+	libraryEngineMinorVersion := libraryEngineSemver.Minor()
+
+	doApiVersionsMatch := libraryEngineMajorVersion == runningEngineMajorVersion && libraryEngineMinorVersion == runningEngineMinorVersion
+	if !doApiVersionsMatch {
+		return nil, stacktrace.NewError(
+			"An API version mismatch was detected between the running engine version '%v' and the engine version the library expects, '%v'; you should use the version of this library that corresponds to the running engine version",
+			runningEngineSemver.String(),
+			libraryEngineSemver.String(),
+		)
+	}
 
 	kurtosisContext := &KurtosisContext{
 		client: engineServiceClient,