+++ engine/api/typescript/src/index.ts
@@ -1,3 +1,6 @@
+// Own Version
+export { KURTOSIS_ENGINE_VERSION } from "./kurtosis_engine_version/kurtosis_engine_version";
+
 export { KurtosisContext, DEFAULT_KURTOSIS_ENGINE_SERVER_PORT_NUM } from "./lib/kurtosis_context/kurtosis_context";
 
 // RPC API bindings