+++ engine/docs/changelog.md
@@ -1,5 +1,15 @@
 # TBD
 
+# 1.8.0
+### Features
+* Added a public constant `KurtosisEngineVersion` in golang and typescript libraries
+* Added a validation to check if the running engine version is the expected in `KurtosisContext` creation
+* Added a user-friendly error text when a client try to create a `KurtosisContext` but the engine is unavailable
+
+### Breaking Changes
+* Renamed the public constant, `DefaultVersion` to `KurtosisEngineVersion` which is more descriptive
+  * Users should replace the constant name in their implementations
+
 # 1.7.7
 ### Fixes
 * Bump to Core 1.36.11, which attempts to pull module & user service images